Move the HTTP request parsing into OpenAPI request manager (#273)
* Move the HTTP request parsing into OpenAPI request manager

* Remove unused imports

* Comment indicator parameter in

---------
